backend/requirements.txt
# Placeholder for requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==0.28.0
google-generativeai==0.3.0
obd==0.7.1
speechrecognition==3.10.0
pyttsx3==2.90
pydub==0.25.1
websockets==12.0
pydantic==2.5.0